requirements.txt
cryptography
pyqrcode
pyzbar
PILLOW
numpy
opencv-python
